+++ tests/tests.rs
@@ -15,19 +15,19 @@ fn char_array()
 fn time_matmul(b: &mut test::Bencher)
 {
     b.iter(|| {
-        let mut a: Array<uint, _> = Array::zeros((2, 3));
+        let mut a: Array<usize, _> = Array::zeros((2, 3));
         for (i, elt) in a.iter_mut().enumerate() {
             *elt = i;
         }
 
-        let mut b: Array<uint, _> = Array::zeros((3, 4));
+        let mut b: Array<usize, _> = Array::zeros((3, 4));
         for (i, elt) in b.iter_mut().enumerate() {
             *elt = i;
         }
 
         let c = a.mat_mul(&b);
         unsafe {
-            let result = Array::from_vec_dim((2, 4), vec![20u, 23, 26, 29, 56, 68, 80, 92]);
+            let result = Array::from_vec_dim((2, 4), vec![20, 23, 26, 29, 56, 68, 80, 92]);
             assert!(c == result);
         }
     })
